Reinstall servicecontrol package on new installer download
This feature addes the ability to reinstall servicecontrol when the remote_file resource downloads a *new version* of service control.  This allows it to act like a real package manager but still needs to be improved to utilize chocolatey in the future for people with more advanced and flexible needs.